SVY-17343 tabIndex doesn't change in NG client
made bstabpanel for TiNG in sync with the default one where
we really test for tab === selectedTab
and not tab.containedForm and tab.relation (because there could be
different tabs that are equals for that)

for some reason the ng1 bs tabpanel already had a very different
implementation in the "select" function.

also fixed the part that when there is a selected tab, we do give the
tab that will be shown already in the hideForm call so no need to call
it again.