SVY-18890 upgrade to angular 17.x and also change the builder to the "application" (esbuild) builder
update to angular 17.1.3 and fix  the accordion to use the new
directives instead of tags.

The one that doesn't work yet is the preventDefault() of the tabchange
and only change the tab on a server call that says ok.
(because the tab is now always switched by the ngbAccordionButton
directive)
see:
https://github.com/ng-bootstrap/ng-bootstrap/issues/4674