SVY-19012 TiNG tabIndex is reverted if there is any other tab model change
if the tabs are again send as changed (right after it is displayed)
and there is no selectedTab yet then don't try to set the selectedtab
because that could override the tabindex that is also coming.

The tab should only be reselected if there was one in the first place if
the tabs property is changed.